Renamed package i18n to internationalization.
