+++ src/main/java/com/mivek/enums/Intensity.java
@@ -1,6 +1,6 @@
 package com.mivek.enums;
 
-import i18n.Messages;
+import internationalization.Messages;
 
 /**
  * Enumeration for indicator.
