+++ src/main/java/internationalization/package-info.java
@@ -2,4 +2,4 @@
  * @author mivek
  *
  */
-package i18n;
+package internationalization;
